+++ src/lib/GameEngine.cs
@@ -121,7 +121,7 @@ namespace IncrementalSociety
 		public bool CanDestoryBuilding (string buildingName) => !ResourceEngine.FindBuilding (buildingName).PreventDestroy;
 		
 		public int GetBuildingTotal (GameState state) => state.AllBuildings ().Count ();
-		public int GetMaxBuildings (GameState state) => PopulationEngine.GetPopUnitsForTotalPopulation (state.Population);
+		public double GetMaxBuildings (GameState state) => PopulationEngine.GetPopUnitsForTotalPopulation (state.Population);
 		public double GetHousingCapacity (GameState state) => PopulationEngine.GetHousingCapacity (state);
 		
 		public bool CanIncreasePopulationCap (GameState state) => PopulationEngine.CanIncreasePopulationCap (state);
